+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * 【core】         Convert.toInt增加容错，NumberUtil增加toNumber方法（issue#IN2LP@Gitee）
 
 ### Bug修复
+* 【core】         修复DateUtil.yearAndQuarter计算错误的问题（issue#IN38V@Gitee）
 
 -------------------------------------------------------------------------------------------------------------